scheduler configure for worker only

+++ config/initializers/cron.rb
@@ -1,9 +1,11 @@
-begin
-  Delayed::Job.where.not(cron: nil).delete_all
-  UpdateStatsJob.set(cron: '45 * * * *').perform_later
-  UpdateAllPodcastsJob.set(cron: '*/10 * * * *').perform_later
-  CleanupLocalMediaJob.set(cron: '3 * * * *').perform_later
-  CleanupOutdatedJob.set(cron: '33 * * * *').perform_later
-rescue
-  Rails.logger.error "Faild to set cron jobs"
+if ENV['SCHEDULER_CONFIGURE'] == 'true'
+  begin
+    Delayed::Job.where.not(cron: nil).delete_all
+    UpdateStatsJob.set(cron: '45 * * * *').perform_later
+    UpdateAllPodcastsJob.set(cron: '*/10 * * * *').perform_later
+    CleanupLocalMediaJob.set(cron: '3 * * * *').perform_later
+    CleanupOutdatedJob.set(cron: '33 * * * *').perform_later
+  rescue
+    Rails.logger.error "Faild to set cron jobs"
+  end
 end
